+++ weaviate/pyproject.toml
@@ -2,7 +2,6 @@
 requires = [
     "hatchling>=0.13.0",
     "setuptools>=66; python_version > '3.0'",
-    "setuptools; python_version < '3.0'",
 ]
 build-backend = "hatchling.build"
 
@@ -11,7 +10,7 @@ name = "datadog-weaviate"
 description = "The Weaviate check"
 readme = "README.md"
 license = {text = "BSD-3-Clause"}
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 keywords = [
     "datadog",
     "datadog agent",
@@ -27,11 +26,11 @@ classifiers = [
     "Intended Audience :: System Administrators",
     "License :: OSI Approved :: BSD License",
     "Private :: Do Not Upload",
-    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
     "Topic :: System :: Monitoring",
 ]
 dependencies = [
-    "datadog-checks-base>=25.1.0",
+    "datadog-checks-base>=32.4.0",
 ]
 dynamic = [
     "version",
