+++ weaviate/tests/test_unit.py
@@ -1,22 +1,16 @@
 # (C) Datadog, Inc. 2023-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-import os
 
 import pytest
 
 from datadog_checks.base.constants import ServiceCheck
-from datadog_checks.dev import get_here
 from datadog_checks.dev.utils import get_metadata_metrics
 from datadog_checks.weaviate import WeaviateCheck
 
-from .common import API_METRICS, MOCKED_INSTANCE, OM_METRICS
+from .common import API_METRICS, MOCKED_INSTANCE, OM_METRICS, get_fixture_path
 
-HERE = get_here()
-
-
-def get_fixture_path(filename):
-    return os.path.join(HERE, 'fixtures', filename)
+pytestmark = pytest.mark.unit
 
 
 @pytest.mark.parametrize(
@@ -32,8 +26,8 @@ def test_check_mock_weaviate_responses(dd_run_check, aggregator, mock_http_respo
     dd_run_check(check)
 
     for metric in metrics:
-        aggregator.assert_metric(metric, at_least=1)
-        aggregator.assert_metric_has_tag(metric, 'test:tag', at_least=1)
+        aggregator.assert_metric(metric)
+        aggregator.assert_metric_has_tag(metric, 'test:tag')
 
     aggregator.assert_all_metrics_covered()
     aggregator.assert_metrics_using_metadata(get_metadata_metrics())
@@ -78,23 +72,3 @@ def test_custom_validation(dd_run_check, instance):
         ):
             check = WeaviateCheck('weaviate', {}, [instance])
             dd_run_check(check)
-
-
-@pytest.mark.integration
-def test_check_mock_weaviate_metadata(dd_run_check, datadog_agent, mock_http_response):
-    mock_http_response(file_path=get_fixture_path('weaviate_meta_api.json'))
-    check = WeaviateCheck('weaviate', {}, [MOCKED_INSTANCE])
-    check.check_id = 'test:123'
-    dd_run_check(check)
-
-    raw_version = '1.19.1'
-    major, minor, patch = raw_version.split('.')
-    version_metadata = {
-        'version.scheme': 'semver',
-        'version.major': major,
-        'version.minor': minor,
-        'version.patch': patch,
-        'version.raw': raw_version,
-    }
-
-    datadog_agent.assert_metadata('test:123', version_metadata)
